Add trio extended (v2) API

+++ python_socks/async_/asyncio/_proxy.py
@@ -47,7 +47,7 @@ class AsyncioProxy(abc.AsyncProxy):
         self,
         dest_host: str,
         dest_port: int,
-        timeout: int = None,
+        timeout: float = None,
         _socket=None,
     ) -> socket.socket:
